Move engines into own module
