+++ docs/api.rst
@@ -5,4 +5,18 @@ API Reference
 ===============
 
 
-.. automodule:: aiodiskqueue
+Queues
+======
+
+.. automodule:: aiodiskqueue.core
+
+Storage Engines
+===============
+
+.. automodule:: aiodiskqueue.engines
+
+
+Exceptions
+==========
+
+.. automodule:: aiodiskqueue.exceptions
