+++ src/aiodiskqueue/__init__.py
@@ -1,6 +1,7 @@
 """Persistent queue for Python AsyncIO."""
 
-from aiodiskqueue.core import PickledList, PickleSequence, Queue
+from aiodiskqueue.core import Queue
+from aiodiskqueue.engines import PickledList, PickleSequence
 from aiodiskqueue.exceptions import QueueEmpty, QueueFull
 
 __version__ = "0.1.0b4"
